+++ scripts/screens/ChangeGenderScreen.py
@@ -14,8 +14,8 @@ from scripts.game_structure.localization import load_lang_resource
 from scripts.game_structure.ui_elements import (
     UIImageButton,
     CatButton,
-    UISurfaceImageButton,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..cat import pronouns
