+++ scripts/screens/ChooseAdoptiveParentScreen.py
@@ -10,8 +10,8 @@ from scripts.game_structure.propagating_thread import PropagatingThread
 from scripts.game_structure.ui_elements import (
     UIImageButton,
     UISpriteButton,
-    UISurfaceImageButton,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
