+++ scripts/screens/ChooseMateScreen.py
@@ -9,8 +9,8 @@ from scripts.game_structure import image_cache
 from scripts.game_structure.ui_elements import (
     UIImageButton,
     UISpriteButton,
-    UISurfaceImageButton,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..events_module.event_filters import get_personality_compatibility
