+++ scripts/screens/ClanSettingsScreen.py
@@ -13,10 +13,10 @@ from scripts.cat.cats import Cat
 from scripts.game_structure.game.settings import game_setting_get
 from scripts.game_structure.ui_elements import (
     UIImageButton,
-    UISurfaceImageButton,
     UIModifiedScrollingContainer,
     UICheckbox,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.housekeeping.datadir import open_data_dir
 from ..ui.theme import get_text_box_theme
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
