+++ scripts/screens/EventEditScreen.py
@@ -21,7 +21,6 @@ from scripts.game_structure import game
 from scripts.cat.pronouns import get_default_pronouns
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
-    UISurfaceImageButton,
     UIModifiedScrollingContainer,
     UITextBoxTweaked,
     UICheckbox,
@@ -31,6 +30,7 @@ from scripts.game_structure.ui_elements import (
     UICollapsibleContainer,
     UIScrollingDropDown,
 )
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.windows.editor_save_check import EditorSaveCheck
 from scripts.ui.windows.editor_missing_info import EditorMissingInfoWindow
 from scripts.screens.RelationshipScreen import RelationshipScreen
