+++ scripts/screens/EventsScreen.py
@@ -21,9 +21,9 @@ from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
     UIModifiedScrollingContainer,
     IDImageButton,
-    UISurfaceImageButton,
     CatButton,
 )
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.screens_core.screens_core import rebuild_moon_n_season_indicator
 from scripts.ui.elements.save_button import UISaveButton
 from scripts.ui.windows.game_over import GameOverWindow
