+++ scripts/screens/ListScreen.py
@@ -23,9 +23,9 @@ from scripts.game_structure.screen_settings import game_screen_size, MANAGER
 from scripts.game_structure.ui_elements import (
     UIImageButton,
     UICatListDisplay,
-    UISurfaceImageButton,
     UIDropDown,
 )
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.Screens import Screens
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import ButtonStyles, get_button_dict
