+++ scripts/screens/MedDenScreen.py
@@ -11,9 +11,9 @@ from scripts.game_structure.ui_elements import (
     UISpriteButton,
     UIImageButton,
     UITextBoxTweaked,
-    UISurfaceImageButton,
     UIModifiedImage,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import event_text_adjust, shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_offset
