+++ scripts/screens/ProfileScreen.py
@@ -17,9 +17,9 @@ from scripts.game_structure import image_cache, game
 from scripts.game_structure.ui_elements import (
     UIImageButton,
     UITextBoxTweaked,
-    UISurfaceImageButton,
     UIModifiedImage,
 )
+from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import (
     process_text,
