+++ scripts/ui/elements/save_button.py
@@ -10,7 +10,7 @@ from scripts.game_structure import image_cache, game
 from scripts.game_structure.game import switch_get_value, Switch
 from scripts.game_structure.game.settings import game_settings_save
 from scripts.game_structure.game.switches import switch_set_value
-from scripts.game_structure.ui_elements import UISurfaceImageButton
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.scale import ui_scale_dimensions, ui_scale
