+++ scripts/ui/windows/confirm_display_change.py
@@ -4,7 +4,7 @@ import pygame
 import pygame_gui
 
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import UISurfaceImageButton
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.window_base_class import GameWindow
 from scripts.ui.scale import ui_scale, ui_scale_offset
