+++ scripts/ui/windows/editor_save_check.py
@@ -3,9 +3,9 @@ import pygame_gui
 import ujson
 
 from scripts.game_structure.ui_elements import (
-    UISurfaceImageButton,
     UITextBoxTweaked,
 )
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.window_base_class import GameWindow
 from scripts.ui.scale import ui_scale
