+++ scripts/ui/windows/no_mediator.py
@@ -7,7 +7,8 @@ from scripts.game_structure.game.switches import (
     Switch,
 )
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import UITextBoxTweaked, UISurfaceImageButton
+from scripts.game_structure.ui_elements import UITextBoxTweaked
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.window_base_class import GameWindow
