+++ scripts/ui/windows/select_focus_clans.py
@@ -4,8 +4,8 @@ import pygame_gui
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
     UIImageButton,
-    UISurfaceImageButton,
 )
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.window_base_class import GameWindow
